feat: replace GeneralUtil popup methods with PopupUtil for better modularity
