--- app/src/main/kotlin/org/db_poultry/controller/MainController.kt
@@ -1,11 +1,0 @@
-package org.db_poultry.controller
-
-import javafx.fxml.FXML
-import javafx.scene.text.Text
-
-class MainController {
-
-    @FXML
-    private lateinit var textSample: Text
-
-}
