+++ app/src/main/kotlin/org/db_poultry/controller/UpdateAddDeleteSuppliesController.kt
@@ -16,6 +16,7 @@ import java.sql.Date
 import javafx.fxml.Initializable
 import javafx.scene.image.Image
 import javafx.scene.image.ImageView
+import org.db_poultry.util.PopupUtil
 import org.db_poultry.util.SupplyTypeSingleton
 import java.io.File
 import java.net.URL
@@ -100,10 +101,10 @@ class UpdateAddDeleteSuppliesController: Initializable {
         val result = createSupplyRecord(getConnection(), supplyID, sqlDate, added, consumed, false)
         if (result != null) {
             undoSingleton.setUndoMode(undoTypes.doUndoSupplyRecord)
-            GeneralUtil.showPopup("success", "Supply record created successfully.")
+            PopupUtil.showPopup("success", "Supply record created successfully.")
             println("Successfully created supply record.")
         } else {
-            GeneralUtil.showPopup("error", "Failed to create supply record.")
+            PopupUtil.showPopup("error", "Failed to create supply record.")
             println("DEBUG: createSupplyRecord returned null")
         }
     }
