+++ app/src/main/kotlin/org/db_poultry/util/GeneralUtil.kt
@@ -10,12 +10,6 @@ import org.kordamp.ikonli.javafx.FontIcon
 
 import javafx.beans.binding.Bindings
 import javafx.beans.property.SimpleDoubleProperty
-import javafx.scene.Scene
-import javafx.scene.paint.Color
-import javafx.scene.text.Text
-import javafx.stage.Modality
-import javafx.stage.Stage
-import javafx.stage.StageStyle
 
 import java.time.format.DateTimeFormatter
 import java.time.LocalDate
@@ -52,88 +46,14 @@ class GeneralUtil {
 
         private var inUseContentPane: AnchorPane? = null
 
-        @JvmStatic
+        fun getInUseContentPane(): AnchorPane? = inUseContentPane
+
         fun registerSectionChangeCallback(callback: (String) -> Unit) {
             sectionChangeCallback = callback
         }
 
-        @JvmStatic
         fun getCurrentSection(): String? = currentSection
 
-        @JvmStatic
-        fun showPopup(status: String, text: String) {
-            val fxmlPath = when (status.lowercase()) {
-                "success" -> "/fxml/popup/popup_success.fxml"
-                "error" -> "/fxml/popup/popup_error.fxml"
-                else -> throw IllegalArgumentException("Invalid status: $status")
-            }
-
-            val loader = FXMLLoader(GeneralUtil::class.java.getResource(fxmlPath))
-            val root = loader.load<Parent>()
-
-            // Set the popup text
-            val textId = if (status.lowercase() == "success") "popupSuccessText" else "popupErrorText"
-            val popupText = root.lookup("#$textId") as? Text
-            popupText?.text = text
-
-            val ownerStage = Stage.getWindows()
-                .filterIsInstance<Stage>()
-                .firstOrNull { it.isFocused }
-
-            val stage = Stage()
-            stage.initStyle(StageStyle.UNDECORATED)
-
-            // Set the owner (this ensures the popup stays on top of the current window)
-            if (ownerStage != null) {
-                stage.initOwner(ownerStage)
-            }
-
-            stage.initModality(Modality.APPLICATION_MODAL)
-
-            // Create scene with transparent background
-            stage.initStyle(StageStyle.TRANSPARENT)
-            val scene = Scene(root)
-            scene.fill = Color.TRANSPARENT
-            stage.scene = scene
-
-            stage.isResizable = false
-
-            // Position the popup when it's shown
-            stage.setOnShown {
-                if (inUseContentPane != null) {
-                    // Get content pane's bounds in screen coordinates
-                    val bounds = inUseContentPane!!.localToScreen(inUseContentPane!!.boundsInLocal)
-
-                    // Calculate center position of the content pane
-                    val centerX = bounds.minX + bounds.width / 2
-                    val centerY = bounds.minY + bounds.height / 2
-
-                    // Position popup window with its center at the content pane's center
-                    stage.x = centerX - stage.width / 2
-                    stage.y = centerY - stage.height / 2
-                } else {
-                    // Fallback to center on owner window if contentPane not provided
-                    stage.centerOnScreen()
-                }
-            }
-
-            var dragOffsetX = 0.0
-            var dragOffsetY = 0.0
-
-            root.setOnMousePressed { event ->
-                dragOffsetX = event.sceneX
-                dragOffsetY = event.sceneY
-            }
-
-            root.setOnMouseDragged { event ->
-                stage.x = event.screenX - dragOffsetX
-                stage.y = event.screenY - dragOffsetY
-            }
-
-            stage.showAndWait()
-        }
-
-        @JvmStatic
         fun loadContentView(contentAnchorPane: AnchorPane, fxmlPath: String) {
             inUseContentPane = contentAnchorPane
 
@@ -187,7 +107,6 @@ class GeneralUtil {
             }
         }
 
-        @JvmStatic
         fun navigateToMainContent(currentPane: AnchorPane, fxmlPath: String) {
 
             val mainContentPane = findMainContentPane(currentPane)
@@ -198,7 +117,6 @@ class GeneralUtil {
             }
         }
 
-        @JvmStatic
         private fun findMainContentPane(currentPane: AnchorPane): AnchorPane? {
             // Traverse up the scene graph to find the main content pane
             var parent = currentPane.parent
@@ -211,15 +129,12 @@ class GeneralUtil {
             return null
         }
 
-
-
         /**
          * Initializes the font size manager for the main pane.
          * This will adjust the font size based on the scene dimensions.
          * mainPane should be the root of your layout.
          * Credits to https://stackoverflow.com/a/51948875
          */
-        @JvmStatic
         fun initializeFontSizeManager(mainPane: Node) {
             mainPane.sceneProperty().addListener { _, oldScene, newScene ->
                 if (oldScene == null && newScene != null) {
@@ -236,7 +151,6 @@ class GeneralUtil {
             }
         }
 
-        @JvmStatic
         fun resizeImageViewToFit(mainNode: Node, toResizeImage: ImageView, widthMultiplier: Double = 0.15, heightMultiplier: Double = 0.2) {
             mainNode.sceneProperty().addListener { _, _, scene ->
                 if (scene != null) {
@@ -246,7 +160,6 @@ class GeneralUtil {
             }
         }
 
-        @JvmStatic
         fun initializeIconSizeManager(mainPane: Node, fontIcon: FontIcon, baseIconSize: Double = 50.0) {
             mainPane.sceneProperty().addListener { _, oldScene, newScene ->
                 if (oldScene == null && newScene != null) {
@@ -262,7 +175,6 @@ class GeneralUtil {
             }
         }
 
-        @JvmStatic
         fun formatDatePretty(date: LocalDate?): String {
             val formatter = DateTimeFormatter.ofPattern("MMMM d, yyyy", Locale.ENGLISH)
             return date?.format(formatter) ?: String.format("No date provided")
